Remove `ConsoleFragmentLogger` (#218)
* Remove `ConsoleFragmentLogger` and add `ConsoleLogger` package trait

* Move `LogRecord` to its own file

* Move `ConsoleLogger` to its own target

* Update README and DocC

* Small fix in DocC

* Move the default `LoggerFragment`s inside an extension

* Revert `ConsoleLoggerExample`

* Enable Windows CI

* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Apply suggestion from @gwynne


* Duplicate `TestConsole`, formatting and typo fixes

* Update Sources/ConsoleLogger/ConsoleLogger.swift


* Update Sources/ConsoleLogger/ConsoleLogger.swift


* Update Sources/ConsoleLogger/ConsoleLogger.swift


* Fix tests and `NonisolatedNonsendingByDefault`

---------
