+++ Package.swift
@@ -10,7 +10,8 @@ let package = Package(
         .tvOS(.v18),
     ],
     products: [
-        .library(name: "ConsoleKit", targets: ["ConsoleKit"])
+        .library(name: "ConsoleKit", targets: ["ConsoleKit"]),
+        .library(name: "ConsoleLogger", targets: ["ConsoleLogger"]),
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-log.git", from: "1.6.3"),
@@ -20,20 +21,37 @@ let package = Package(
         .target(
             name: "ConsoleKit",
             dependencies: [
-                .product(name: "Logging", package: "swift-log"),
-                .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
+                .product(name: "AsyncAlgorithms", package: "swift-async-algorithms")
             ],
             swiftSettings: swiftSettings
         ),
         .testTarget(
             name: "ConsoleKitTests",
-            dependencies: [.target(name: "ConsoleKit")],
+            dependencies: [
+                .target(name: "ConsoleKit")
+            ],
+            swiftSettings: swiftSettings
+        ),
+        .target(
+            name: "ConsoleLogger",
+            dependencies: [
+                .product(name: "Logging", package: "swift-log"),
+                .target(name: "ConsoleKit"),
+            ],
+            swiftSettings: swiftSettings
+        ),
+        .testTarget(
+            name: "ConsoleLoggerTests",
+            dependencies: [
+                .target(name: "ConsoleLogger")
+            ],
             swiftSettings: swiftSettings
         ),
         .executableTarget(
             name: "ConsoleLoggerExample",
             dependencies: [
                 .target(name: "ConsoleKit"),
+                .target(name: "ConsoleLogger"),
                 .product(name: "Logging", package: "swift-log"),
             ],
             swiftSettings: swiftSettings
@@ -45,6 +63,11 @@ var swiftSettings: [SwiftSetting] {
     [
         .enableUpcomingFeature("ExistentialAny"),
         .enableUpcomingFeature("MemberImportVisibility"),
+        .enableUpcomingFeature("ExistentialAny"),
         .enableUpcomingFeature("InternalImportsByDefault"),
+        .enableUpcomingFeature("MemberImportVisibility"),
+        .enableUpcomingFeature("InferIsolatedConformances"),
+        //.enableUpcomingFeature("NonisolatedNonsendingByDefault"),
+        .enableUpcomingFeature("ImmutableWeakCaptures"),
     ]
 }
