+++ Sources/ConsoleLogger/Docs.docc/index.md
@@ -1,0 +1,39 @@
+# ``ConsoleLogger``
+
+@Metadata {
+    @TitleHeading(Package)
+}
+
+A `SwiftLog` `LogHandler` implementation for customizable logging to a console.
+
+## Overview
+
+`ConsoleLogger` is a flexible logging backend for console applications, allowing developers to customize log output with various fragments, including timestamps, log levels, and source locations.
+
+## Topics
+
+### Essentials
+
+- ``ConsoleLogger``
+- ``LogRecord``
+- ``ConsoleLogger/Logging``
+
+### Logger Fragments
+
+- ``LoggerFragment``
+- ``FragmentOutput``
+- ``IfMaxLevelFragment``
+- ``AndFragment``
+- ``LabelFragment``
+- ``LevelFragment``
+- ``LiteralFragment``
+- ``SeparatorFragment``
+- ``MessageFragment``
+- ``MetadataFragment``
+- ``SourceLocationFragment``
+- ``LoggerSourceFragment``
+- ``TimestampSource``
+- ``SystemTimestampSource``
+- ``TimestampFragment``
+- ``DefaultLoggerFragmentType``
+- ``TimestampDefaultLoggerFragmentType``
