+++ Sources/ConsoleLogger/LoggerFragments/LoggerFragment.swift
@@ -1,3 +1,4 @@
+public import ConsoleKit
 public import Logging
 
 #if canImport(Darwin)
@@ -13,69 +14,6 @@ public import WASILibc
 #elseif os(Windows)
 public import CRT
 #endif
-
-/// Information about a specific log message, including information from the logger the message was logged to.
-public struct LogRecord {
-    public init(
-        level: Logger.Level,
-        message: Logger.Message,
-        metadata: Logger.Metadata? = nil,
-        source: String,
-        file: String,
-        function: String,
-        line: UInt,
-        label: String,
-        loggerLevel: Logger.Level,
-        loggerMetadata: Logger.Metadata,
-        metadataProvider: Logger.MetadataProvider? = nil
-    ) {
-        self.level = level
-        self.message = message
-        self.metadata = metadata
-        self.source = source
-        self.file = file
-        self.function = function
-        self.line = line
-        self.label = label
-        self.loggerLevel = loggerLevel
-        self.loggerMetadata = loggerMetadata
-        self.metadataProvider = metadataProvider
-    }
-
-    /// The log level of the message
-    public var level: Logger.Level
-    /// The logged message
-    public var message: Logger.Message
-    /// The metadata explicitly associated with the logged message
-    public var metadata: Logger.Metadata?
-    /// The source of the log message, usually the module name
-    public var source: String
-    /// The file the message was logged from
-    public var file: String
-    /// The function the message was logged from
-    public var function: String
-    /// The line number in the file the message was logged from
-    public var line: UInt
-
-    /// The label of the logger the message was logged to
-    public var label: String
-    /// The log level of the logger the message was logged to
-    public var loggerLevel: Logger.Level
-    /// The metadata associated with the logger the message was logged to
-    public var loggerMetadata: Logger.Metadata
-    /// The metadata provider associated with the logger the message was logged to
-    public var metadataProvider: Logger.MetadataProvider?
-
-    /// Combine all of the metadata into a single set.
-    public mutating func allMetadata() -> [String: Logger.MetadataValue] {
-        // We aren't mutating self here currently,
-        // but keeping the method marked that way will ensure we can cache the result
-        // without breaking the public API if we decide that's desirable.
-        (self.metadata ?? [:])
-            .merging(self.loggerMetadata, uniquingKeysWith: { (a, _) in a })
-            .merging(self.metadataProvider?.get() ?? [:], uniquingKeysWith: { (a, _) in a })
-    }
-}
 
 /// The output of a ``LoggerFragment``, including some intermediary state used for things like deduplicating separators.
 public struct FragmentOutput {
@@ -368,17 +306,24 @@ extension Logger.MetadataValue {
         case .array(let array): return "[\(array.map(\.descriptionWithoutExcessQuotes).joined(separator: ", "))]"
         case .dictionary(let dict): return "[\(dict.map { "\($0): \($1.descriptionWithoutExcessQuotes)" }.joined(separator: ", "))]"
         case .string(let str): return str
-        case .stringConvertible(let conv): return "\(conv)"
+        case .stringConvertible(let repr):
+            return switch repr {
+            case let repr as Bool: "\(repr)"
+            case let repr as any FixedWidthInteger: "\(repr)"
+            case let repr as any BinaryFloatingPoint: "\(repr)"
+            default: #""\#(repr.description)""#
+            }
         }
     }
 }
 
 extension Logger.Metadata {
     fileprivate var sortedDescriptionWithoutQuotes: String {
-        let contents = Array(self)
-            .sorted(by: { $0.0 < $1.0 })
+        "["
+            + self.lazy
+            .sorted(by: { $0.key < $1.key })
             .map { "\($0): \($1.descriptionWithoutExcessQuotes)" }
             .joined(separator: ", ")
-        return "[\(contents)]"
+            + "]"
     }
 }
