+++ Tests/ConsoleLoggerTests/LoggingTests.swift
@@ -1,4 +1,5 @@
 import ConsoleKit
+import ConsoleLogger
 import Logging
 import Testing
 
@@ -90,7 +91,8 @@ struct ConsoleLoggerTests {
             ]
         )
         expect(
-            console, logs: .info, message: #"info [meta1: test1, meta2: Missing command, meta3: [hello, wor"ld], meta4: [hello: wor"ld]]"#)
+            console, logs: .info, message: #"info [meta1: test1, meta2: "Missing command", meta3: [hello, wor"ld], meta4: [hello: wor"ld]]"#
+        )
     }
 
     @Test("Source Location")
@@ -101,7 +103,7 @@ struct ConsoleLoggerTests {
         }
 
         logger.debug("debug", line: 1)
-        expect(console, logs: .debug, message: "debug (ConsoleKitTests/LoggingTests.swift:1)")
+        expect(console, logs: .debug, message: "debug (ConsoleLoggerTests/LoggingTests.swift:1)")
     }
 
     @Test("Metadata Providers")
@@ -121,7 +123,7 @@ struct ConsoleLoggerTests {
         TraceNamespace.$simpleTraceID.withValue("1234-5678") {
             logger.debug("debug", line: 1)
         }
-        expect(console, logs: .debug, message: "debug [simple-trace-id: 1234-5678] (ConsoleKitTests/LoggingTests.swift:1)")
+        expect(console, logs: .debug, message: "debug [simple-trace-id: 1234-5678] (ConsoleLoggerTests/LoggingTests.swift:1)")
     }
 
     @Test("Timestamp Fragment")
@@ -145,8 +147,8 @@ struct ConsoleLoggerTests {
             time.tm_mon = 5
             time.tm_year = 100
 
-            return ConsoleFragmentLogger(
-                fragment: timestampDefaultLoggerFragment(timestampSource: ConstantTimestampSource(time: time)),
+            return ConsoleLogger(
+                fragment: .timestampDefault(timestampSource: ConstantTimestampSource(time: time)),
                 label: label,
                 console: console
             )
@@ -162,7 +164,7 @@ struct ConsoleLoggerTests {
         // Remove the timezone, since there doesn't appear to be a good way to mock it with strftime.
         while logged.removeFirst() != " " {}
 
-        #expect(logged == "[ \(Logger.Level.info.name) ] logged (ConsoleKitTests/LoggingTests.swift:1)\n")
+        #expect(logged == "[ \(Logger.Level.info.name) ] logged (ConsoleLoggerTests/LoggingTests.swift:1)\n")
     }
 
     @Test("Source Fragment")
@@ -170,8 +172,8 @@ struct ConsoleLoggerTests {
         let console = TestConsole()
 
         let logger = Logger(label: "codes.vapor.console") { label in
-            ConsoleFragmentLogger(
-                fragment: LoggerSourceFragment().and(defaultLoggerFragment().separated(" ")),
+            ConsoleLogger(
+                fragment: LoggerSourceFragment().and(.default.separated(" ")),
                 label: label,
                 console: console
             )
@@ -180,7 +182,8 @@ struct ConsoleLoggerTests {
         logger.info("logged", line: 1)
 
         #expect(
-            console.testOutputQueue.first == "ConsoleKitTests [ \(Logger.Level.info.name) ] logged (ConsoleKitTests/LoggingTests.swift:1)\n"
+            console.testOutputQueue.first
+                == "ConsoleLoggerTests [ \(Logger.Level.info.name) ] logged (ConsoleLoggerTests/LoggingTests.swift:1)\n"
         )
     }
 }
